Agregado endpoint de bankAccount
